Move SQL queries into a dedicated namespace
