+++ src/Akeneo/Pim/Structure/Family/back/Test/Integration/Query/SqlCountFamilyCodesIntegration.php
@@ -4,7 +4,7 @@ namespace Akeneo\Test\Pim\Structure\Family\Integration\Query;
 
 use Akeneo\Pim\Structure\Family\ServiceAPI\Query\FamilyQuery;
 use Akeneo\Pim\Structure\Family\ServiceAPI\Query\CountFamilyCodes;
-use Akeneo\Pim\Structure\Family\Infrastructure\Query\SqlCountFamilyCodes;
+use Akeneo\Pim\Structure\Family\Infrastructure\Query\Sql\SqlCountFamilyCodes;
 use Akeneo\Pim\Structure\Family\ServiceAPI\Query\FamilyQuerySearch;
 use Akeneo\Test\Integration\Configuration;
 use Akeneo\Test\Integration\TestCase;
