+++ src/Akeneo/Pim/Structure/Family/back/Test/Integration/Query/SqlFindFamilyCodesIntegration.php
@@ -6,7 +6,7 @@ use Akeneo\Pim\Structure\Family\ServiceAPI\Query\FamilyQuery;
 use Akeneo\Pim\Structure\Family\ServiceAPI\Query\FamilyQueryPagination;
 use Akeneo\Pim\Structure\Family\ServiceAPI\Query\FamilyQuerySearch;
 use Akeneo\Pim\Structure\Family\ServiceAPI\Query\FindFamilyCodes;
-use Akeneo\Pim\Structure\Family\Infrastructure\Query\SqlFindFamilyCodes;
+use Akeneo\Pim\Structure\Family\Infrastructure\Query\Sql\SqlFindFamilyCodes;
 use Akeneo\Test\Integration\Configuration;
 use Akeneo\Test\Integration\TestCase;
 use Webmozart\Assert\Assert;
